🛂 Also add way to automatically create admin user via config
